feat: modularity part 1
