+++ src/crypto/__init__.py
@@ -1,0 +1,6 @@
+"""Cryptographic modules for the Fractum application."""
+
+from src.crypto.memory import SecureMemory, SecureContext
+from src.crypto.encryption import FileEncryptor
+
+__all__ = ['SecureMemory', 'SecureContext', 'FileEncryptor']
