+++ src/utils/__init__.py
@@ -1,0 +1,5 @@
+"""Utility functions for the Fractum application."""
+
+from src.utils.integrity import calculate_tool_integrity, get_enhanced_random_bytes
+
+__all__ = ['calculate_tool_integrity', 'get_enhanced_random_bytes']
